Functional regarding the recent additions
ADD: The content of the non-wanted numbers is now functional.
ADD: The statistic at the end are now correctly computed and displayed. There is also statistic regarding the draw that have been discarded since the non-wanted number list is now taken in account.
ADD: Improvement in the presentation of the summary at the end.
ADD: Add "Mise-Éclair" function to have application suggest numbers.